config/Base: Inline `test()` fn
