+++ src/tests/util/test_app.rs
@@ -1,8 +1,8 @@
 use super::{MockAnonymousUser, MockCookieUser, MockTokenUser};
 use crate::record;
 use crate::util::{chaosproxy::ChaosProxy, fresh_schema::FreshSchema};
-use crates_io::config::{self, BalanceCapacityConfig, DatabasePools, DbPoolConfig};
-use crates_io::{background_jobs::Environment, env, App, Emails};
+use crates_io::config::{self, BalanceCapacityConfig, Base, DatabasePools, DbPoolConfig};
+use crates_io::{background_jobs::Environment, env, App, Emails, Env, Uploader};
 use crates_io_index::testing::UpstreamIndex;
 use crates_io_index::{Credentials, Repository as WorkerRepository, RepositoryConfig};
 use std::{rc::Rc, sync::Arc, time::Duration};
@@ -333,6 +333,33 @@ impl TestAppBuilder {
 }
 
 fn simple_config() -> config::Server {
+    let uploader = Uploader::S3 {
+        bucket: Box::new(s3::Bucket::new(
+            dotenvy::var("TEST_S3_BUCKET").unwrap_or_else(|_err| "crates-test".into()),
+            parse_test_region(dotenvy::var("TEST_S3_REGION").ok()),
+            dotenvy::var("TEST_AWS_ACCESS_KEY").unwrap_or_default(),
+            dotenvy::var("TEST_AWS_SECRET_KEY").unwrap_or_default(),
+            // When testing we route all API traffic over HTTP so we can
+            // sniff/record it, but everywhere else we use https
+            "http",
+        )),
+        index_bucket: Some(Box::new(s3::Bucket::new(
+            dotenvy::var("TEST_S3_INDEX_BUCKET").unwrap_or_else(|_err| "crates-index-test".into()),
+            parse_test_region(dotenvy::var("TEST_S3_INDEX_REGION").ok()),
+            dotenvy::var("TEST_AWS_ACCESS_KEY").unwrap_or_default(),
+            dotenvy::var("TEST_AWS_SECRET_KEY").unwrap_or_default(),
+            // When testing we route all API traffic over HTTP so we can
+            // sniff/record it, but everywhere else we use https
+            "http",
+        ))),
+        cdn: None,
+    };
+
+    let base = Base {
+        env: Env::Test,
+        uploader,
+    };
+
     let db = DatabasePools {
         primary: DbPoolConfig {
             url: env("TEST_DATABASE_URL").into(),
@@ -346,7 +373,7 @@ fn simple_config() -> config::Server {
     };
 
     config::Server {
-        base: config::Base::test(),
+        base,
         db,
         session_key: cookie::Key::derive_from("test this has to be over 32 bytes long".as_bytes()),
         gh_client_id: ClientId::new(dotenvy::var("GH_CLIENT_ID").unwrap_or_default()),
@@ -376,6 +403,21 @@ fn simple_config() -> config::Server {
     }
 }
 
+static DEFAULT_TEST_REGION: &str = "127.0.0.1:19000";
+
+fn parse_test_region(maybe_region: Option<String>) -> s3::Region {
+    match maybe_region {
+        Some(region) if region.contains("://") => {
+            let (_proto, host) = region.split_once("://").unwrap();
+            s3::Region::Host(host.to_string())
+        }
+        Some(region) if !region.is_empty() => s3::Region::Region(region),
+        // An empty or missing region will use the default. This needs to match the region
+        // configuration that was used to generate the cassettes in `src/tests/http-data`.
+        _ => s3::Region::Host(DEFAULT_TEST_REGION.to_string()),
+    }
+}
+
 fn build_app(config: config::Server, proxy: Option<String>) -> (Arc<App>, axum::Router) {
     let client = if let Some(proxy) = proxy {
         let mut builder = Client::builder();
@@ -399,4 +441,20 @@ fn build_app(config: config::Server, proxy: Option<String>) -> (Arc<App>, axum::
     let app = Arc::new(app);
     let router = crates_io::build_handler(Arc::clone(&app));
     (app, router)
+}
+
+#[test]
+fn test_parse_region() {
+    for (input, expected) in [
+        (None, s3::Region::Host(DEFAULT_TEST_REGION.into())),
+        (Some(""), s3::Region::Host(DEFAULT_TEST_REGION.into())),
+        (Some("us-west-2"), s3::Region::Region("us-west-2".into())),
+        (Some("http://foo.bar"), s3::Region::Host("foo.bar".into())),
+        (
+            Some("https://127.0.0.1:9000"),
+            s3::Region::Host("127.0.0.1:9000".into()),
+        ),
+    ] {
+        assert_eq!(parse_test_region(input.map(String::from)), expected);
+    }
 }
